Make replaceable_objects.py a module; split out stored procedures
